Refactor elements and add simple action handler
