--- src/main/java/noob/gui/elements/UserInput.java
@@ -1,9 +1,0 @@
-package noob.gui.elements;
-
-import javafx.scene.control.TextField;
-
-public class UserInput extends TextField {
-    public void setStyle(int width) {
-        this.setPrefWidth(width);
-    }
-}
